fix status function on setting page
